Complete all sections and the project as a whole
In the initial edits, I updated the header section to include two flex
boxes, so the header links stay in a proper position relative to the
whole webpage. Specifically, I matched the in-line structure based on
the goal image. Due to this change, the styling was redistributed to the
new flex box containers. Afterwards, I work on the content and styling
of the last three sections: quote, call to action, and footer. Now, I
have completed the project. Cheers to my second Odin Project. :)